+++ src/components/Callout/Callout.stories.tsx
@@ -1,14 +1,14 @@
-import { Callout, CalloutProps } from "./Callout";
+import { Callout, CalloutProps } from './Callout';
 
 export default {
-  title: "Components/Callout",
+  title: 'Components/Callout',
   component: Callout,
 };
 
 export const Default = {
   render: (args: CalloutProps) => <Callout {...args} />,
   args: {
-    variant: "error",
-    message: "Diese Aktion konnte nicht ausgeführt werden.",
+    variant: 'error',
+    message: 'Diese Aktion konnte nicht ausgeführt werden.',
   },
 };