+++ src/components/Layout/Pages/ContentWithHeader/DefaultWithHeader.stories.tsx
@@ -1,11 +1,11 @@
-import { Meta, StoryObj } from "@storybook/react";
-import { ContentWithHeader } from "./ContentWithHeader";
+import { Meta, StoryObj } from '@storybook/react';
+import { ContentWithHeader } from './ContentWithHeader';
 
 const meta: Meta<typeof ContentWithHeader> = {
-  title: "Layout/Pages/ContentWithHeader",
+  title: 'Layout/Pages/ContentWithHeader',
   component: ContentWithHeader,
   parameters: {
-    layout: "fullscreen",
+    layout: 'fullscreen',
   },
   argTypes: {
     Header: {
@@ -24,9 +24,9 @@ export default meta;
 export const Default: StoryObj<typeof ContentWithHeader> = {
   render: (args) => <ContentWithHeader Header={args.Header} Main={args.Main} />,
   args: {
-    Header: <div style={{ backgroundColor: "green" }}>Header</div>,
+    Header: <div style={{ backgroundColor: 'green' }}>Header</div>,
     Main: (
-      <div style={{ backgroundColor: "lightgreen", height: "1000px" }}>
+      <div style={{ backgroundColor: 'lightgreen', height: '1000px' }}>
         Scrollable Main Content
       </div>
     ),