+++ src/legacy/Button/index.ts
@@ -1,0 +1,3 @@
+export { Button as LegacyButton } from './Button';
+export { ButtonGroup as LegacyButtonGroup } from './ButtonGroup';
+export { IconButton as LegacyIconButton } from './IconButton';